Spin off PrefetchedRange.h
